pyproject.toml
[tool.poetry]
name = "Metakam"
version = "0.1.0"
description = ""
authors = ["YevheniiMelnikov <evgen.dev@metakam.ua>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^74.1.2"
psycopg2-binary = "^2.9"
django = "^5.1.1"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.2"
djangorestframework-api-key = "^3.0.0"
drf-yasg = "^1.21.7"
djoser = "^2.2.3"
django-filter = "^24.3"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.3"
mypy = "^1.11.2"
mypy-extensions = "^1.0.0"

[tool.ruff]
line-length = 120
exclude = [
  ".git",
  ".mypy_cache",
  ".pytest_cache",
  ".env",
  "env",
  "venv",
  ".venv",
  "./metakam/api/migrations/",
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.env
    | env
    | venv
    | .venv
    | metakam/api/migrations
  )
)
'''

[tool.mypy]
follow_imports = "skip"
strict_optional = false
check_untyped_defs = false
allow_untyped_globals = true
ignore_missing_imports = true
disallow_untyped_calls = true
warn_unused_ignores = true
warn_unreachable = true
exclude = '''
(
  \.git
| \.mypy_cache
| \.pytest_cache
| \.env
| env
| .venv
| venv
| metakam/api/migrations
)
'''
